1. Fix bugs in Plus operation of EV+MDD. A value + Infinity = Infinity.
2. Fix bugs in constraint saturation with EV+MDD.

git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/cjiang-ev@768 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/meddly_expert.h
@@ -510,7 +510,8 @@ class MEDDLY::unpacked_node {
     static unpacked_node* newFromNode(const expert_forest *f, node_handle node, bool full);
 
     static unpacked_node* newRedundant(const expert_forest *f, int k, node_handle node, bool full);
-    static unpacked_node* newRedundant(const expert_forest *f, int k, int ev, node_handle node, bool full);
+//    static unpacked_node* newRedundant(const expert_forest *f, int k, int ev, node_handle node, bool full);
+    static unpacked_node* newRedundant(const expert_forest *f, int k, long ev, node_handle node, bool full);
     static unpacked_node* newRedundant(const expert_forest *f, int k, float ev, node_handle node, bool full);
 
     static unpacked_node* newIdentity(const expert_forest *f, int k, int i, node_handle node, bool full);
